refactor(perf): mixture of server and client loading
add caching to some api calls
