+++ src/lib/server/repo/projectRepo.js
@@ -17,15 +17,19 @@ export async function getProjectsWithCategories(term, start, end, supabase) {
     .from('projects')
     .select(
       `
-      *,
-      category_project (
-        categories (
-          *
+      id,
+      title,
+      banner_image,
+      funding_goal,
+      current_funding,
+      user_id,
+      category_project!inner (
+        categories!inner (
+          image
         )
       )
     `,
     )
-    .ilike('title', `%${term}%`)
     .range(start, end)
     .order('created_at', { ascending: false });
