+++ src/routes/api/projects/+server.js
@@ -6,7 +6,7 @@ import { getProjectsWithDetails } from '$lib/server/service/projectService.js';
 
 import { json } from '@sveltejs/kit';
 
-export async function GET({ url, locals }) {
+export async function GET({ url, locals, setHeaders }) {
   const page = parseInt(url.searchParams.get('page') || '1', 10);
   const limit = parseInt(url.searchParams.get('limit') || '6', 10);
   const term = url.searchParams.get('term') || '';
@@ -14,6 +14,11 @@ export async function GET({ url, locals }) {
 
   try {
     const projects = await getProjectsWithDetails(term, page, limit, supabase);
+
+    setHeaders({
+      'Cache-Control': 'public, max-age=600, stale-while-revalidate=300',
+      Vary: 'Accept-Encoding',
+    });
 
     return json({ projects: projects }, { status: 200 });
   } catch (error) {
